@@ -3,8 +3,8 @@
     "type": "phpbb-extension",
     "description": "Only serves as a temporary solution until phpBB itself offers the appropriate settings: This will disable the S9E TextFormatter Emoji plugin for new or modified posts. In addition, the display of Emojis in existing posts is suppressed. This also applies to PMs.",
     "homepage": "https://github.com/LukeWCS/disable-emojis",
-    "version": "1.1.0",
-    "time": "2023-12-09",
+    "version": "1.1.1",
+    "time": "2024-12-11",
     "license": "GPL-2.0-only",
     "authors": [
         {
@@ -14,13 +14,13 @@
         }
     ],
     "require": {
-        "php": ">=7.1.0,<8.4.0@dev",
+        "php": ">=7.4.0,<8.5.0@dev",
         "composer/installers": "~1.0.0"
     },
     "extra": {
         "display-name": "Disable Emojis",
         "soft-require": {
-            "phpbb/phpbb": ">=3.2.11,<3.4.0@dev"
+            "phpbb/phpbb": ">=3.3.0,<3.4.0@dev"
         },
         "version-check": {
             "host": "lukewcs.github.io",
