@@ -12,15 +12,15 @@ namespace lukewcs\disableemojis\controller;
 
 class acp_disableemojis_controller
 {
-	protected $language;
-	protected $template;
-	protected $request;
-	protected $config;
-	protected $cache;
-	protected $ext_manager;
+	protected object $language;
+	protected object $template;
+	protected object $request;
+	protected object $config;
+	protected object $cache;
+	protected object $ext_manager;
 
-	protected $metadata;
-	public $u_action;
+	public    string $u_action;
+	protected array  $metadata;
 
 	public function __construct(
 		\phpbb\language\language $language,
@@ -44,7 +44,7 @@ class acp_disableemojis_controller
 	public function module_settings()
 	{
 		$this->language->add_lang(['acp_disableemojis'], 'lukewcs/disableemojis');
-		$this->set_meta_template_vars('DISABLEEMOJIS');
+		$this->set_meta_template_vars('DISABLEEMOJIS', 'LukeWCS');
 
 		if ($this->request->is_set_post('submit'))
 		{
@@ -52,24 +52,26 @@ class acp_disableemojis_controller
 			{
 				trigger_error($this->language->lang('FORM_INVALID') . adm_back_link($this->u_action), E_USER_WARNING);
 			}
+
 			$delete_cache = ($this->request->variable('disableemojis_save_emoji_token', 0) != $this->config['disableemojis_save_emoji_token']);
 			$this->config->set('disableemojis_save_emoji_token'		, $this->request->variable('disableemojis_save_emoji_token', 0));
 			$this->config->set('disableemojis_replace_token_mode'	, $this->request->variable('disableemojis_replace_token_mode', 0));
+
 			if ($delete_cache)
 			{
 				$this->cache->purge();
 			}
+
 			trigger_error($this->language->lang('DISABLEEMOJIS_MSG_SAVED_SETTINGS') . adm_back_link($this->u_action));
 		}
 
 		$this->template->assign_vars([
 			'DISABLEEMOJIS_SAVE_EMOJI_TOKEN'				=> $this->config['disableemojis_save_emoji_token'],
-			'DISABLEEMOJIS_REPLACE_TOKEN_MODE'				=> $this->config['disableemojis_replace_token_mode'],
-			'DISABLEEMOJIS_REPLACE_TOKEN_MODE_OPTIONS' => [
+			'DISABLEEMOJIS_REPLACE_TOKEN_MODE_OPTIONS'		=> $this->select_struct($this->config['disableemojis_replace_token_mode'], [
 				'DISABLEEMOJIS_REPLACE_TOKEN_DO_NOTHING' 	=> '0',
 				'DISABLEEMOJIS_REPLACE_TOKEN_KEEP_CODE' 	=> '1',
 				'DISABLEEMOJIS_REPLACE_TOKEN_SHOW_HINT' 	=> '2',
-			],
+			]),
 			'U_ACTION'										=> $this->u_action,
 		]);
 
@@ -81,19 +83,41 @@ class acp_disableemojis_controller
 		$this->u_action = $u_action;
 	}
 
-	private function set_meta_template_vars(string $tpl_prefix): void
+	private function set_meta_template_vars(string $tpl_prefix, string $copyright): void
 	{
-		$this->template->assign_vars([
-			$tpl_prefix . '_METADATA'	=> [
-				'EXT_NAME'		=> $this->metadata['extra']['display-name'],
-				'EXT_VER'		=> $this->language->lang($tpl_prefix . '_VERSION_STRING', $this->metadata['version']),
-			] + ($this->language->is_set($tpl_prefix . '_LANG_DESC') ? [
-				'LANG_DESC'		=> $this->language->lang($tpl_prefix . '_LANG_DESC'),
-				'LANG_VER'		=> $this->language->lang($tpl_prefix . '_VERSION_STRING', $this->language->lang($tpl_prefix . '_LANG_VER')),
-				'LANG_AUTHOR'	=> $this->language->lang($tpl_prefix . '_LANG_AUTHOR'),
-			] : []) + [
-				'CLASS'			=> strtolower($tpl_prefix) . '_footer',
-			],
-		]);
+		$template_vars = [
+			'ext_name'		=> $this->metadata['extra']['display-name'],
+			'ext_ver'		=> $this->language->lang($tpl_prefix . '_VERSION_STRING', $this->metadata['version']),
+			'ext_copyright'	=> $copyright,
+			'class'			=> strtolower($tpl_prefix) . '_footer',
+		];
+		$template_vars += $this->language->is_set($tpl_prefix . '_LANG_VER') ? [
+			'lang_desc'		=> $this->language->lang($tpl_prefix . '_LANG_DESC'),
+			'lang_ver'		=> $this->language->lang($tpl_prefix . '_VERSION_STRING', $this->language->lang($tpl_prefix . '_LANG_VER')),
+			'lang_author'	=> $this->language->lang($tpl_prefix . '_LANG_AUTHOR'),
+		] : [];
+
+		$this->template->assign_vars([$tpl_prefix . '_METADATA' => $template_vars]);
+	}
+
+	private function select_struct($cfg_value, array $options): array
+	{
+		$options_tpl = [];
+
+		foreach ($options as $opt_key => $opt_value)
+		{
+			if (!is_array($opt_value))
+			{
+				$opt_value = [$opt_value];
+			}
+			$options_tpl[] = [
+				'label'		=> $opt_key,
+				'value'		=> $opt_value[0],
+				'bold'		=> $opt_value[1] ?? false,
+				'selected'	=> is_array($cfg_value) ? in_array($opt_value[0], $cfg_value) : $opt_value[0] == $cfg_value,
+			];
+		}
+
+		return $options_tpl;
 	}
 }
